feat: Add --output flag to context-jetpack
The context-jetpack script now supports an optional --output flag. This allows users to specify a directory where the extracted Jetpack source code will be placed. If the flag is not provided, the script will continue to create a temporary directory. The usage documentation has been updated to reflect this new option.